+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-02 19:37:31
+
+### Changed Files
+- `src/features/hierarchy/components/AgentTable.tsx`
+- `src/services/policies/PolicyRepository.ts`
+- `supabase/functions/slack-policy-notification/index.ts`
+
+### Statistics
+```
+ 3 files changed, 37 insertions(+), 11 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-02 19:17:48
 
 ### Changed Files